Add files via upload
models.py file is from an app called surveyorportal and in that file am interested in the observation file.
rinextest.py is a code to read the observation file and its in the media directory. For this script, i need it to iterate over the uploaded observation files and for each write the results into a text file.